feat: add per-backend opt-out build tags for the cross-platform backends
Adds a keyring_no<backend> build-tag family so consumers can exclude
backends they never use (and their dependency trees) at build time:

- keyring_no1password: op / op-connect / op-desktop (drops
  onepassword-sdk-go incl. wazero, and connect-sdk-go incl.
  jaeger-client-go)
- keyring_nofile: file (drops dvsekhvalnov/jose2go)
- keyring_nopass: pass
- keyring_nopassage: passage

Default (untagged) builds are unaffected: constants and Config fields
stay in ungated files, the package import graph is unchanged, and an
excluded backend behaves exactly like an unavailable platform backend
(absent from AvailableBackends, ErrNoAvailImpl from Open).

Test files are gated to match their sources; runCmd moves to a neutral
cmdtestutil_test.go because passage_test.go shares it with pass_test.go.
mocks_test.go gets its constraint via mockery template-data
(mock-build-tags) so regeneration preserves it; this also fixes the
latent freebsd test-compile breakage (op sources are !freebsd but their
tests were unconstrained). A new CI leg builds, vets, and runs the
gated regression test with all four tags; together with the default
build it brackets every tag combination, since each tag only removes
whole files and no gated file references another tag group's symbols.

Implements https://github.com/ByteNess/keyring/issues/93
