+++ .github/workflows/test.yaml
@@ -58,3 +58,18 @@ jobs:
           go-version-file: 'go.mod'
           check-latest: true
       - run: go test -race ./...
+
+  # The default build above and the all-tags build below bracket every tag
+  # combination: each opt-out tag only removes whole files, and no gated file
+  # references a symbol from a different tag group.
+  opt-out-tags:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v6
+      - uses: actions/setup-go@v6
+        with:
+          go-version-file: 'go.mod'
+          check-latest: true
+      - run: go build -tags keyring_no1password,keyring_nofile,keyring_nopass,keyring_nopassage ./...
+      - run: go vet -all -tags keyring_no1password,keyring_nofile,keyring_nopass,keyring_nopassage ./...
+      - run: go test -race -run 'TestOptOutTagsExcludeBackends' -tags keyring_no1password,keyring_nofile,keyring_nopass,keyring_nopassage ./...
