+++ opstandard.go
@@ -1,4 +1,4 @@
-//go:build !freebsd
+//go:build !freebsd && !keyring_no1password
 
 package keyring
