+++ optestutil_test.go
@@ -1,3 +1,5 @@
+//go:build !freebsd && !keyring_no1password
+
 package keyring
 
 import (
