+++ pass.go
@@ -1,5 +1,4 @@
-//go:build !windows
-// +build !windows
+//go:build !windows && !keyring_nopass
 
 package keyring
