+++ passage_test.go
@@ -1,4 +1,4 @@
-//go:build !windows
+//go:build !windows && !keyring_nopassage
 
 package keyring
